requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1
httpx>=0.27,<1
python-dotenv>=1.0,<2
pytest>=8.0,<9
pytest-httpx>=0.30,<1
